feat(editor): surface all line/shape properties in the properties panel
The annotation properties panel only showed Color and Line type for a
selected /Line or /PolyLine. It now also shows the stroke width and
opacity sliders (the restyle path already supported both) and start/end
line-ending dropdowns, so lines and shapes expose their full editable
style there — matching the toolbar's style menu.

Extract the toolbar's line-ending preview painter and label into a shared
line_ending_controls.dart (PdfLineEndingDropdown / PdfLineEndingPainter /
pdfLineEndingLabel) so the panel and toolbar render the same control
without duplicating the painter.
